rbppc_nand: Clean up and bring up to date with modern kernel APIs.
These changes allow rbppc_nand to have a partition layout configured using a
device-tree blob, or, optionally, command-line parameters. It removes support
for a static partition layout entirely. Currently, the partition layout for the
RB333 and RB600 has been migrated to their respective device-tree blob entries.

All global state has been removed, and the driver now looks for any node that is
compatible with "rb,nand". This means that this module could theoretically work
with multiple NAND devices if someone felt like soldering an extra few onto the
board and figuring out where the GPIO is for it (unlikely, I know). Mostly I
don't like global state.

The driver is updated to use the devres APIs, which means we don't have to worry
so much about freeing resources (the previous iteration of the driver allocated
plenty but freed little). It also correctly reserves memory correctly before
ioremap()ing, although it's arguably not important to do so.