pata_rbppc_upm: Partial rewrite of pata_rbppc_cf driver.
This rewrite does not change any functionality provided by the pata_rbppc_cf
driver, with the exception of being able to reprogram the UPM for PIO modes
other than 0 (previously, the driver would only move to lower PIO modes and the
ATA code always requested PIO mode 0 first).

The inspiration for this rewrite was the removal of rb_iomap.c, upon which this
module formerly depended. The entirety of rb_iomap.c has been replaced by a
single macro (REG_OFFSET()) in this module.

I also took the liberty of cleaning up the code to generate timings. It's still
not great, but it looks better than it did before. There are still a few
improvements that could be made around setting the PIO mode.