rbppc_gtm_beeper: Add support for GTM-based beeper.
Rather than simply provide an exported-but-unused rbppc_beep() function, this
driver correctly registers sound support with the input device libraries in the
kernel; these can be exposed to userland through evdev.

This driver (of course) uses the proper kernel APIs for accessing the GTM and
GPIO functionality required for the RB333, rather than the hacked-together
ioremap()s and low-level IO used in the old beeper initialization/event code
that was part of the rbppc.c platform.