powerpc: Add support for the RouterBOARD 333 and 600 series devices.
The RouterBOARD 333 and 600 series are Freescale MPC832x and MPC834x SOC-
based embedded systems (respectively). They are gifted with a terrible device
tree, and use built-in DTBs for configuration here. Custom platform boot code
is used for device-specific initialization.

This addition is accompanied by three RouterBOARD-specific device drivers:

* pata_rbppc_upm: This driver provides access to the CompactFlash card slots
  built into the board. It's accessed via the Freescale UPM, which is
  programmed according to the slowest device currently attached to the
  machine using PIO only.

* rbppc_nand: This is a GPIO-based NAND flash driver. It bears some
  resemblance to the generic GPIO driver, but uses a custom ECC layout and
  has a few other quirks. It may be a candidate for merging with the generic
  GPIO driver at a later point.

* rbppc_gtm_beeper: A simple driver that provides access to the on-board
  beeper device through the GTM on the Freescale chip.

All of these changes are based on the kernel patches released under the GPL
by Mikrotikls SIA, the manufacturer of the RouterBOARD devices.

Signed-off-by: Noah Fontes <nfontes@invectorate.com>